Remove deprecations
