+++ typings/DisTube.d.ts
@@ -364,13 +364,6 @@ declare class DisTube extends EventEmitter {
      */
     private _isVoiceChannelEmpty;
     /**
-     * TODO: Remove this
-     * @deprecated use {@link DisTube#addRelatedVideo} instead
-     * @param {DisTube.Message} message Message
-     * @returns {Promise<Queue>}
-     */
-    runAutoplay(message: any): Promise<Queue>;
-    /**
      * Add related song to the queue
      * @async
      * @param {Discord.Snowflake|Discord.Message} message The message from guild channel
