--- typings/duration.d.ts
@@ -1,2 +1,0 @@
-export function formatDuration(milliseconds: any): string;
-export function toSecond(string: any): number;
