+++ typings/util.d.ts
@@ -1,0 +1,3 @@
+export function formatDuration(milliseconds: number): string;
+export function toSecond(string: string): number;
+export function parseNumber(string: string): number;
